requirements.txt
fastapi
uvicorn
matplotlib
pillow
streamlit
pandas
plotly
requests
